Fix crash in cache_artwork() with invalid SD keys errorCode and errorMessage
Sometimes the sd4tvh add-on crashes when retrieving Schedules Direct
listings.  The crash seems to occur when the listing metadata
indicates episode artwork is available, but the artwork image is
missing.

The root cause appears to be that the SD keys "errorCode" and
"errorMessage" are not valid.

Fix by using the SD keys "code" and "message" respectively.

BACKGROUND INFORMATION:

The following link is the closest I've found to documentation
describing the key names for errors:

    https://github.com/SchedulesDirect/JSON-Service/wiki/API-20141201#errors

Following is an example of the crash error message in the sd4tvh.log
file:

Traceback (most recent call last):
  File "/storage/.kodi/addons/script.module.sd4tvh/sd4tvh.py", line 407, in <module>
    main()
  File "/storage/.kodi/addons/script.module.sd4tvh/sd4tvh.py", line 403, in main
    app.process()
  File "/storage/.kodi/addons/script.module.sd4tvh/sd4tvh.py", line 114, in process
    self._sd.refresh_cache(schedule_hash_list)
  File "/storage/.kodi/addons/script.module.sd4tvh/libschedulesdirect/schedulesdirect.py", line 177, in refresh_cache
    self.cache_artwork()
  File "/storage/.kodi/addons/script.module.sd4tvh/libschedulesdirect/schedulesdirect.py", line 163, in cache_artwork
    self._logger.warn(u"Artwork for %s returned %s %s", artwork_error["programID"], artwork_error["data"]["errorCode"], artwork_error["data"]["errorMessage"])
KeyError: 'errorCode'

After applying this enhancement, the warning that caused the crash can
be seen in the sd4tvh.log file, such as this example snippet:

WARNING:libschedulesdirect.schedulesdirect:Artwork for EP00188710 returned 5000 Could not find requested image. Post message to http://forums.schedulesdirect.org/viewforum.php?f=6 if you are having issues.